\name{families.ergmm}
\alias{families.ergmm}
\alias{ergmm.families}
\alias{dlpY.deta.Bernoulli.logit}
\alias{dlpY.deta.binomial.logit}
\alias{dlpY.deta.fs}
\alias{dlpY.deta.Poisson.log}
\alias{lpY.Bernoulli.logit}
\alias{lpY.binomial.logit}
\alias{lpYc.Bernoulli.logit}
\alias{lpYc.binomial.logit}
\alias{lpYc.fs}
\alias{lpYc.Poisson.log}
\alias{lpY.Poisson.log}
\alias{lpY.fs}
\alias{EY.Bernoulli.logit}
\alias{EY.binomial.logit}
\alias{EY.fs}
\alias{EY.Poisson.log}
\alias{mk.lp.optim.fs}
\alias{pY.fs}
\alias{pY.Poisson.log}
\alias{rsm.fs}
\alias{rsm.Poisson.log}
\alias{rsm.binomial.logit}
\alias{rsm.Bernoulli.logit}
\alias{pY.Bernoulli.logit}
\alias{pY.binomial.logit}
\alias{family.IDs}
\alias{family.names}
\alias{fam.par.check}
\title{Edge Weight Distribution Families}
\description{
  Family-link combinations supported by \code{\link{ergmm}}.
}
\section{Family-link combinations}{
  \tabular{rlll}{
    ID \tab C name          \tab Family    \tab Link  \cr
     0 \tab Bernoulli\_logit \tab Bernoulli \tab logit \cr
     1 \tab binomial\_logit  \tab binomial  \tab logit \cr
     2 \tab Poisson\_log     \tab Possion   \tab log   \cr
     3 \tab Bernoulli\_logit \tab Bernoulli \tab logit \cr
     4 \tab Bernoulli\_logit \tab Bernoulli \tab logit
  }
}
