\name{latentnet-internal}
\title{Internal latentnet Objects}

\alias{DistMatrix}
\alias{InitErgm.latentcov}
\alias{InitErgm.latent}
\alias{InitErgm.latentcluster}
\alias{ergm.Cprepare.latent}
\alias{ergm.getmodel.latent}
\alias{ergm.getterms.latent}
\alias{ergm.gettermnames.latent}
\alias{ergm.latent}
\alias{ergm.statseval.latent}
\alias{ergm.latentplot}
\alias{mlpY.grad.plot}
\alias{mlpY.plot}
\alias{mlpYmdsZ.grad.plot}
\alias{mlpYmdsZ.plot}
%
\alias{ergm.geodesicmatrix}
\alias{ergm.geodesicmatrix.edgelist}
\alias{ergm.latentcluster}
\alias{ergm.MCMCacf}
\alias{ergm.procAdj}
\alias{ergm.procAdj.fcnt}
\alias{ergm.statseval.latentcluster}
\alias{is.latent}
\alias{is.latent.cluster}
\alias{klswitch.c}
\alias{latent.wrapper}
\alias{lpz.dist}
\alias{mcmc.diagnostics.default}
\alias{mlpY}
\alias{mlpY0}
\alias{mlpY0.grad}
\alias{mlpY.grad}
\alias{mlpYmdsZ}
\alias{mlpYmdsZ.grad}
\alias{mlpYnull}
\alias{mvimode}
\alias{permutation}
\alias{plot.mcmc.ergm}
\alias{print.raftery.diag.ergm}
\alias{rergm.latent}
\alias{rergm.latent.sociomatrix}
\alias{set.mfrow}
\alias{traceplot.ergm}
\alias{add.contours}
\alias{drawcircle}
\alias{drawpie}
\alias{labelswitch}
\alias{midarrow}
\alias{ergm.plinfo.latent}
\alias{network.dyadcount}
\alias{robust.inverse}
%
% These need to be documented!
%
\alias{summary.ergm}
\alias{summary.statsmatrix.ergm}
\alias{is.invertible}

\description{
  Internal latentnet functions.
}
%\usage{
%}
%\details{
%  Most of these are not to be called by the user.
%}
%\arguments{
% \item{\dots}{Type help(package="latentnet") to get started.}
%}
\seealso{ergm}
\keyword{internal}
