\name{families.ergmm}
\alias{families.ergmm}
\alias{ergmm.families}
\title{Edge Weight Distribution Families}
\description{
  Family-link combinations supported by \code{\link{ergmm}}.
}
\section{Family-link combinations}{
  \tabular{rlll}{
    ID \tab C name          \tab Family    \tab Link  \cr
     0 \tab Bernoulli_logit \tab Bernoulli \tab logit \cr
     1 \tab binomial_logit  \tab binomial  \tab logit \cr
     2 \tab Poisson_log     \tab Possion   \tab log   \cr
     3 \tab Bernoulli_logit \tab Bernoulli \tab logit \cr
     4 \tab Bernoulli_logit \tab Bernoulli \tab logit
  }
}