\name{latentnet-internal}
\title{Internal latentnet Objects}

%
% These might be worth documenting
%

\alias{mbc.VII.EM}

\alias{print.ergmm.model}
\alias{ergmm.model}
\alias{ergmm.model.object}

\alias{getYm}
\alias{bipartite.augment}
\alias{observed.dyads}

\alias{statsreeval.ergmm}

\alias{ergmm.lpY}
\alias{ergmm.lpY.C}
\alias{ergmm.lpY.grad}
\alias{ergmm.lp}
\alias{ergmm.lpBeta}
\alias{ergmm.lpBeta.grad}
\alias{ergmm.lp.grad}
\alias{ergmm.lp.grad.approx}
\alias{ergmm.lpLV}
\alias{ergmm.lpLV.grad}
\alias{ergmm.lpRE}
\alias{ergmm.lpRE.grad}
\alias{ergmm.lpREV}
\alias{ergmm.lpREV.grad}
\alias{ergmm.lpZ}
\alias{ergmm.lpZ.grad}
\alias{ergmm.lpdispersion}
\alias{ergmm.lpdispersion.grad}
\alias{ergmm.eta}
\alias{ergmm.EY}
\alias{lp.works}
\alias{lpsum}

\alias{cov.beta.ext}
\alias{get.init.deltas}
\alias{get.sample.deltas}

\alias{find.mle}
\alias{find.mle.loop}
\alias{find.mpe}
\alias{find.pmode.loop}
\alias{find.mkl}

\alias{add.mcmc.mle.mle.ergmm}
\alias{add.mcmc.pmode.pmode.ergmm}
\alias{add.mkl.mbc.ergmm}
\alias{add.mkl.pos.ergmm}

\alias{xtabs.ergmm}
\alias{clust.homogeneity}

%% \alias{procr}
%% \alias{procr.ergmm.model}
%% \alias{procr.matrix}
%% \alias{procr.sample.ergmm}
%% \alias{procrustes.Z.mean.C}

\description{
  Internal latentnet functions.
}
%\usage{
%}
%\details{
%  Most of these are not to be called by the user.
%}
%\arguments{
% \item{\dots}{Type help(package="latentnet") to get started.}
%}
\seealso{ergmm}
\keyword{internal}
