\name{latentnet-internal}
\title{Internal latentnet Objects}

%
% Genuinely internal
%

%
% These need to be documented!
%
\alias{summary.statsmatrix.ergmm}
\alias{is.invertible}

\description{
  Internal latentnet functions.
}
%\usage{
%}
%\details{
%  Most of these are not to be called by the user.
%}
%\arguments{
% \item{\dots}{Type help(package="latentnet") to get started.}
%}
\seealso{ergmm}
\keyword{internal}
